correct bug

+++ src/main/java/org/jibe77/hermanas/data/repository/SensorRepository.java
@@ -3,6 +3,6 @@ package org.jibe77.hermanas.data.repository;
 import org.jibe77.hermanas.data.entity.Sensor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SensorRepository extends CrudRepository<Sensor, Integer> {
+public interface SensorRepository extends CrudRepository<Sensor, Long> {
 
 }
